feat(wp8): pack bulk ADC samples to 24-bit, 4096 -> 6144 samples
The ADS131M04 codes are 24-bit; int32 storage wasted one byte per
channel on sign extension. Store them packed (3 bytes/channel,
little-endian signed) so the same ~50%-of-SRAM budget holds 50% more
samples.

- config.h: ADC_BULK_SAMPLE_COUNT 4096 -> 6144, ADC_BULK_CHUNK_SAMPLES
  7 -> 10, new ADC_BULK_BYTES_PER_SAMPLE 12. Buffer 64 KiB -> 72 KiB
  (RAM 72% -> 77.6%). Capture span ~197 ms -> ~295 ms.
- svc_signal_analysis: s_cap_buf is now a flat uint8_t[N*12]; on_sample
  packs each channel as 3 LE bytes. capture_buffer() returns const
  uint8_t*.
- svc_api bulk_pump: chunk payload is [page][sample:12]xN (was x16),
  10 samples/chunk (was 7).
- apiv2.py: ADC_BULK_* constants, decode_bulk_adc_chunk unpacks 3-byte
  signed LE. bulk_adc_csv.py / adc_diag.py unchanged (go through those).

Bench (v0.8.10, UART): 6144 samples / 295 ms / 0 drops / 20827 Hz.
Round-trips correctly (DC levels and RMS match the v0.8.9 int32 capture);
A/B still peak at 2604 Hz, S1/S2 at the noise floor.

v0.8.9 -> v0.8.10

+++ Services/svc_signal_analysis.h
@@ -47,7 +47,7 @@ bool      svc_signal_analysis_is_running(void);
 DrvStatus      svc_signal_analysis_capture_begin(void);
 bool           svc_signal_analysis_capture_done(void);
 void           svc_signal_analysis_capture_end(void);
-const int32_t *svc_signal_analysis_capture_buffer(void);   /* [count*4] flattened, ch0..ch3 interleaved */
+const uint8_t *svc_signal_analysis_capture_buffer(void);   /* count * ADC_BULK_BYTES_PER_SAMPLE bytes: per sample, ch0..ch3 as 3-byte LE signed */
 uint16_t       svc_signal_analysis_capture_sample_count(void);
 uint16_t       svc_signal_analysis_capture_drops(void);    /* trigger ticks skipped (DRDY not ready) during the fill */
